==> ConvolutionMedicalNer/04-Apply-med7-on-Clinical-Notes.ipynb <==
import pandas as pd
import spacy

med7 = spacy.load("en_core_med7_lg")

preprocessed_df = pd.read_pickle("data/preprocessed_notes.p")

preprocessed_df['ner'] = None

count = 0
preprocessed_index = {}
n = preprocessed_df.shape[0] #181483

for i in preprocessed_df.itertuples():
    
    if count % 1000 == 0:
        print(count)
    if count > 0 and count % 10000 == 0:
        filepath = '../../drive/MyDrive/GT/data/ner_df_{c}.p'.format(c=count)
        pd.to_pickle(preprocessed_df, filepath)

    count += 1
    ind = i.Index
    text = i.preprocessed_text
    
    all_pred = []
    for each_sent in text:
        try:
            doc = med7(each_sent)
            result = ([(ent.text, ent.label_) for ent in doc.ents])
            if len(result) == 0: continue
            all_pred.append(result)
        except:
            print("error..")
            continue
    preprocessed_df.at[ind, 'ner'] = all_pred

pd.to_pickle(preprocessed_df, "data/ner_df.p")
